moving workspace switcher to right side
- changed workspace icon to Ps icon
- added icon and jsx to right side
- removed from left side

review fixes

- white space nits
- new line nit
